+++ packages/auth-services/src/auth-server/server.ts
@@ -1,8 +1,9 @@
-import { env } from '../_setup/env';
+import { env } from '../env';
 import { createLitAuthServer } from './src/createAuthServer';
 
 const litAuthServer = createLitAuthServer({
-  port: env.PORT,
+  port: env.AUTH_SERVER_PORT,
+  host: env.AUTH_SERVER_HOST,
   network: env.NETWORK,
   litTxsenderRpcUrl: env.LIT_TXSENDER_RPC_URL,
   litTxsenderPrivateKey: env.LIT_TXSENDER_PRIVATE_KEY,