+++ packages/auth-services/src/login-server/server.ts
@@ -1,21 +1,21 @@
+import { env } from '../env';
 import { createLitLoginServer } from './src';
 
 // Create a new server instance
 const litLoginServer = createLitLoginServer({
-  // Optional configuration
-  port: 3300,
-  host: '0.0.0.0',
-  stateExpirySeconds: 30,
+  port: env.LOGIN_SERVER_PORT,
+  host: env.LOGIN_SERVER_HOST,
+  stateExpirySeconds: env.LOGIN_SERVER_STATE_EXPIRY_SECONDS,
 
   // Required credentials
   socialProviders: {
     google: {
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+      clientId: env.LOGIN_SERVER_GOOGLE_CLIENT_ID,
+      clientSecret: env.LOGIN_SERVER_GOOGLE_CLIENT_SECRET,
     },
     discord: {
-      clientId: process.env.DISCORD_CLIENT_ID!,
-      clientSecret: process.env.DISCORD_CLIENT_SECRET!,
+      clientId: env.LOGIN_SERVER_DISCORD_CLIENT_ID,
+      clientSecret: env.LOGIN_SERVER_DISCORD_CLIENT_SECRET,
     },
   },
 });