+++ packages/auth-services/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lit-protocol/auth-services
 
+## 1.0.0-alpha.0
+
+### Major Changes
+
+- merged lit-login-server & relayer as auth-services
+
+### Patch Changes
+
+- Updated dependencies
+  - @lit-protocol/constants@8.0.0-alpha.5
+
 ## 0.1.0
 
-- Initial package setup for PKP Auth Service. 
+- Initial package setup for PKP Auth Service.