+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lit-protocol/auth
 
+## 8.0.0-alpha.5
+
+### Major Changes
+
+- merged lit-login-server & relayer as auth-services
+
 ## 8.0.0-alpha.4
 
 ### Major Changes